FM2-APP: Make UI display friendly when DUT is in landscape mode
There is a favorite list in FM2 which contain 7 buttons to save
favorite channel. When DUT is in landscape, the buttons do not fill
the screen in landscape(there is a black blank in the right),
so the UI becomes unfriendly.

To fix the issue, set each button's weight with 1 to make
the buttons with same width and set horizontal scroll view's
attribute fillViewport to true so that it stretches the content
to fill the available width

Change-Id: I290dc0b86b44de1b6405bc2bbee423d6e6e886ed
CRs-Fixed: 492461